fix: rename react directory to reactjs
the directory being named "react" causes an error with the moduleDirectories option from jest
