+++ package.json
@@ -60,7 +60,7 @@
     "createAsyncStoragePersistor-experimental",
     "broadcastQueryClient-experimental",
     "lib",
-    "react",
+    "reactjs",
     "scripts",
     "types"
   ],
@@ -81,7 +81,11 @@
     }
   },
   "typesVersions": {
-    "<4.1": { "types/*": ["types/ts3.8/*"] }
+    "<4.1": {
+      "types/*": [
+        "types/ts3.8/*"
+      ]
+    }
   },
   "devDependencies": {
     "@babel/cli": "^7.11.6",
