+++ src/core/tests/hydration.test.tsx
@@ -1,4 +1,4 @@
-import { mockNavigatorOnLine, sleep } from '../../react/tests/utils'
+import { mockNavigatorOnLine, sleep } from '../../reactjs/tests/utils'
 import { QueryCache } from '../queryCache'
 import { QueryClient } from '../queryClient'
 import { dehydrate, hydrate } from '../hydration'
