+++ src/reactjs/reactBatchedUpdates.native.ts
