+++ src/reactjs/reactBatchedUpdates.ts
