+++ src/reactjs/setBatchUpdatesFn.ts
