+++ src/reactjs/setLogger.ts
