+++ src/reactjs/useBaseQuery.ts
